[feat] Tune Explorers page styles
