+++ aim/web/ui/src/components/kit_v2/Text/Text.d.ts
@@ -11,6 +11,8 @@ export interface ITextProps
   disabled?: boolean;
   size?: CSS['fontSize'];
   color?: CSS['color'];
+  truncate?: boolean;
+  textTransform?: CSS['textTransform'];
   css?: CSS;
 }
