+++ aim/web/ui/src/components/kit_v2/Text/Text.style.ts
@@ -4,6 +4,15 @@ import { styled } from 'config/stitches';
 
 const StyledSlot: any = styled(Slot, {
   lineHeight: 1,
+  variants: {
+    truncate: {
+      true: {
+        overflow: 'hidden',
+        textOverflow: 'ellipsis',
+        whiteSpace: 'nowrap',
+      },
+    },
+  },
 });
 
 export { StyledSlot };
