+++ aim/web/ui/src/components/kit_v2/Text/Text.tsx
@@ -24,6 +24,8 @@ const Text = React.forwardRef<React.ElementRef<typeof StyledSlot>, ITextProps>(
       weight = '$2',
       color = '$textPrimary',
       disabled = false,
+      textTransform,
+      truncate,
       css,
       children,
       ...rest
@@ -33,11 +35,13 @@ const Text = React.forwardRef<React.ElementRef<typeof StyledSlot>, ITextProps>(
     const TagElement = as;
     return (
       <StyledSlot
+        truncate={truncate}
         css={{
           fontSize: size,
           fontWeight: weight,
           color: disabled ? `${color}50` : color,
           fontFamily: mono ? '$mono' : '$inter',
+          textTransform,
           ...css,
         }}
         ref={forwardedRef}
